requirements-dev.txt
pytest-pokie
pytest==8.3.4
pytest-cov==4.0.0
pytest-testdox==3.0.1
flake8==6.0.0
flake8_black==0.3.6
coverage==7.2.5
tox==4.5.1
tox-docker==5.0.0
